==> lib/H_pyseer_all/pyseer_paloc/3_collect_data_from_all_model_runs.R <==
# Goal: collect R-squared value from all models generated by "2_write_sbat.R"

library(tidyverse)
source("2_write_sbat.R")

num_models <- num_n_fold * num_alpha * num_lineage * num_genotype * num_phenotype

model_results <- as.data.frame(matrix(NA, ncol = 7, nrow = num_models))
colnames(model_results) <- c("n_fold", 
                             "alpha",
                             "lineage", 
                             "genotype", 
                             "phenotype",
                             "model_name",
                             "rsq")
row_num <- 0
for (i in 1:num_n_fold) {
  for (j in 1:num_alpha) {
    for (k in 1:num_lineage) {
      for (l in 1:num_genotype) {
        for (m in 1:num_phenotype) {
            row_num <- row_num + 1
            rsquared <- "TBD"
            model_name <- paste0("EN_risk_score_", n_folds_num[i], "_fold_", 
                                 alpha_value_vec[j], "_alpha_", 
                                 names(lineage_vec)[k], "_", 
                                 names(genotype_path_vec)[l], "_", 
                                 names(pheno_paths)[m])
            
            model_out_name <- paste0(model_name, ".out")
            model_results_name <- paste0("../../../data/15_pyseer/paloc_model_results/",  model_name, ".txt")
            model_out <- readLines(model_out_name)
            rsquared <- as.numeric(gsub("Best R.*: ", "",  model_out[grepl("Best R", model_out)]))
          
            if (is_empty(rsquared)) {
              rsquared <- "error"
            }
            
            model_results[row_num, ] <- 
              c(n_folds_num[i], 
                alpha_value_vec[j], 
                names(lineage_vec)[k], 
                names(genotype_path_vec)[l], 
                names(pheno_paths)[m],
                model_name, 
                rsquared) 
        }
      }
    }
  }
}

write_tsv(model_results, path = "../../../data/15_pyseer/paloc_model_results/EN_model_summary.tsv", col_names = TRUE)
